Add 'bdd_load'
The LibBDD package has no support for creating a BDD bottom-up level by level.
But, in the cases it matters we do so from a file LibBDD often runs out of
memory. But, for those benchmarks we may just load the file directly into
LibBDD anyway (since that is where it came from).